Add a col layer to an arphit plot.

Add a col layer to an arphit plot.

agg_col(aes = NULL, data = NULL, colour = NULL, barcol = NULL,
  panel = "1", stacked = TRUE, color)

Arguments

aes
The aesthetic that defines the layer. Will inherit (or parts thereof) if omitted.

data
The data to be used. Will inherit from parent if missing.

colour
A colour to be applied to all of the series, or (if your aesthetic has a group), a vector of colours that will be cycled through to consecutive group elements.

barcol
(optional) Outline colours for each bar series

panel
(default = "1") Which panel of the graph to place this layer on. You can specify a vector of panels (e.g. `panel = c("1","3")`) to apply the layer to multiple panels at once.

stacked
(default = TRUE) Stack the bars, or group them?

color
(Deprecated; use colour instead) A colour to be applied to all of the series, or (if your aesthetic has a group), a vector of colours that will be cycled through to consecutive group elements.

See also

Examples

data  <- data.frame(unemployment = rnorm(20), state = c(rep("A", 10), rep("B", 10)),
  date = seq.Date(from = as.Date("2017-01-10"), length.out = 10, by = "quarter"))
arphitgg(data) + agg_col(aes = agg_aes(x = date, y = unemployment, group = state), panel = "1")

#> dev.new(): using pdf(file="Rplots112.pdf")